frontend: refactored containers, added login
container, configured auth guard
